Focus problem and Multi to radios
When you click on a show towards the right it brings you all the way back left.  Now after it is done performing the action it focuses back on the correct button.
Also, on the multi adder you could choose more than one status.  Changed the check boxes to radio buttons and problem solved.