Fixed 2 HUGE bugs
Bug 1 was uncovered by Dustin when he could not get to 12/7/97.  That
was me not deleting the show tags when a tag was deleted.  Also
rectified the production database.
Bug 2 was not filtering tags on the Notes page by user.  Are you kidding
me, this is why you need a QA dept.